requirements.txt
asgiref==3.5.1
attrs==21.4.0
backports.zoneinfo;python_version<"3.9"
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.4
django-extensions==3.1.5
django-on-heroku==1.1.2
extensions==0.4
Faker==13.11.0
fpdf2==2.5.5
gunicorn==20.1.0
iniconfig==1.1.1
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
python-dateutil==2.8.2
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
whitenoise==6.2.0

pip~=23.1.2
wheel~=0.37.1
setuptools~=65.3.0
djangorestframework~=3.13.1